requirements.txt
accelerate==1.5.2
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrdict==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
contourpy==1.3.0
cycler==0.12.1
deepseek-vl==0.0.1.dev0
deepseek_vl2==1.0.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
latex2mathml==3.77.0
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdtex2html==1.3.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==1.68.2
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==10.4.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.11.2
safetensors==0.5.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
timm==1.0.15
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.38.2
triton==2.0.0
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==12.0
xformers==0.0.22
zipp==3.21.0
zstandard==0.23.0
